Adiciona arquivos de configuração e testes iniciais, incluindo .gitignore, conftest.py, e requirements.txt; ajusta métodos na classe Matricula e documenta a classe Turma.

+++ models/matricula.py
@@ -63,7 +63,7 @@ class Matricula:
         """Indica se a matrícula ainda está ativa (não concluída e não trancada)."""
         return self._ativa
 
-    def lancar_avaliacao(self, nota: float, frequencia: float):
+    def lancar_avaliacao(self, nota: float, frequencia: int):
         """
         Registra nota e frequência do aluno e atualiza automaticamente sua situação.
